Move flipper icons to the layout layer
