+++ frontend/src/layouts/MainLayout.vue
@@ -248,6 +248,23 @@ import * as flipper from '../flipper/core'
 import asyncSleep from 'simple-async-sleep'
 import log from 'loglevel'
 
+const flipperIcons = {
+  'archive:new': 'img:icons/flipper/action-new.svg',
+  'archive:remove': 'img:icons/flipper/action-remove.svg',
+  'archive:rename': 'img:icons/flipper/action-rename.svg',
+  'archive:save': 'img:icons/flipper/action-save.svg',
+  'archive:sdcard': 'img:icons/flipper/location-sdcard.svg',
+  'archive:internal': 'img:icons/flipper/location-internal.svg',
+  'archive:file': 'img:icons/flipper/file.svg',
+  'archive:folder': 'img:icons/flipper/folder.svg',
+  'archive:badusb': 'img:icons/flipper/badusb.svg',
+  'archive:ibutton': 'img:icons/flipper/ibutton.svg',
+  'archive:infrared': 'img:icons/flipper/infrared.svg',
+  'archive:nfc': 'img:icons/flipper/nfc.svg',
+  'archive:rfid': 'img:icons/flipper/rfid.svg',
+  'archive:subghz': 'img:icons/flipper/subghz.svg',
+  'archive:u2f': 'img:icons/flipper/u2f.svg'
+}
 let dismissNotif
 
 export default defineComponent({
@@ -260,6 +277,13 @@ export default defineComponent({
 
   setup () {
     const $q = useQuasar()
+
+    $q.iconMapFn = (iconName) => {
+      const icon = flipperIcons[iconName]
+      if (icon !== void 0) {
+        return { icon: icon }
+      }
+    }
     return {
       routes: [
         {
@@ -279,7 +303,7 @@ export default defineComponent({
         },
         {
           title: 'CLI',
-          icon: 'terminal',
+          icon: 'mdi-console-line',
           link: '/cli'
         },
         {
