+++ frontend/src/pages/Archive.vue
@@ -295,26 +295,9 @@
 
 <script>
 import { defineComponent, ref } from 'vue'
-import { exportFile, useQuasar } from 'quasar'
+import { exportFile } from 'quasar'
 import ProgressBar from 'components/ProgressBar.vue'
 import asyncSleep from 'simple-async-sleep'
-const flipperIcons = {
-  'archive:new': 'img:icons/flipper/action-new.svg',
-  'archive:remove': 'img:icons/flipper/action-remove.svg',
-  'archive:rename': 'img:icons/flipper/action-rename.svg',
-  'archive:save': 'img:icons/flipper/action-save.svg',
-  'archive:sdcard': 'img:icons/flipper/location-sdcard.svg',
-  'archive:internal': 'img:icons/flipper/location-internal.svg',
-  'archive:file': 'img:icons/flipper/file.svg',
-  'archive:folder': 'img:icons/flipper/folder.svg',
-  'archive:badusb': 'img:icons/flipper/badusb.svg',
-  'archive:ibutton': 'img:icons/flipper/ibutton.svg',
-  'archive:infrared': 'img:icons/flipper/infrared.svg',
-  'archive:nfc': 'img:icons/flipper/nfc.svg',
-  'archive:rfid': 'img:icons/flipper/rfid.svg',
-  'archive:subghz': 'img:icons/flipper/subghz.svg',
-  'archive:u2f': 'img:icons/flipper/u2f.svg'
-}
 
 export default defineComponent({
   name: 'PageArchive',
@@ -330,14 +313,6 @@ export default defineComponent({
   },
 
   setup () {
-    const $q = useQuasar()
-
-    $q.iconMapFn = (iconName) => {
-      const icon = flipperIcons[iconName]
-      if (icon !== void 0) {
-        return { icon: icon }
-      }
-    }
     return {
       path: ref('/'),
       dir: ref([]),
